Reorganize and simplify FailureRecoveryTests
