+++ plugin/trino-delta-lake/src/test/java/io/trino/plugin/deltalake/BaseDeltaFailureRecoveryTest.java
@@ -11,14 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.faulttolerant.iceberg;
+package io.trino.plugin.deltalake;
 
-import io.trino.faulttolerant.BaseFailureRecoveryTest;
 import io.trino.operator.RetryPolicy;
 import io.trino.spi.ErrorType;
+import io.trino.testing.BaseFailureRecoveryTest;
 import org.testng.annotations.Test;
 
-import java.util.List;
 import java.util.Optional;
 
 import static io.trino.execution.FailureInjector.FAILURE_INJECTION_MESSAGE;
@@ -27,12 +26,12 @@ import static io.trino.execution.FailureInjector.InjectedFailureType.TASK_GET_RE
 import static io.trino.execution.FailureInjector.InjectedFailureType.TASK_GET_RESULTS_REQUEST_TIMEOUT;
 import static io.trino.execution.FailureInjector.InjectedFailureType.TASK_MANAGEMENT_REQUEST_FAILURE;
 import static io.trino.execution.FailureInjector.InjectedFailureType.TASK_MANAGEMENT_REQUEST_TIMEOUT;
-import static java.lang.String.format;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-public abstract class BaseIcebergFailureRecoveryTest
+public abstract class BaseDeltaFailureRecoveryTest
         extends BaseFailureRecoveryTest
 {
-    protected BaseIcebergFailureRecoveryTest(RetryPolicy retryPolicy)
+    protected BaseDeltaFailureRecoveryTest(RetryPolicy retryPolicy)
     {
         super(retryPolicy);
     }
@@ -43,27 +42,6 @@ public abstract class BaseIcebergFailureRecoveryTest
         return true;
     }
 
-    @Override
-    protected void createPartitionedLineitemTable(String tableName, List<String> columns, String partitionColumn)
-    {
-        String sql = format(
-                "CREATE TABLE %s WITH (partitioning=array['%s']) AS SELECT %s FROM tpch.tiny.lineitem",
-                tableName,
-                partitionColumn,
-                String.join(",", columns));
-        getQueryRunner().execute(sql);
-    }
-
-    @Test(invocationCount = INVOCATION_COUNT)
-    public void testCreatePartitionedTable()
-    {
-        testTableModification(
-                Optional.empty(),
-                "CREATE TABLE <table> WITH (partitioning = ARRAY['p']) AS SELECT *, 'partition1' p FROM orders",
-                Optional.of("DROP TABLE <table>"));
-    }
-
-    // Copied from BaseDeltaFailureRecoveryTest
     @Override
     public void testDelete()
     {
@@ -105,6 +83,14 @@ public abstract class BaseIcebergFailureRecoveryTest
                 .failsWithoutRetries(failure -> failure.hasMessageContaining(FAILURE_INJECTION_MESSAGE))
                 .finishesSuccessfully();
 
+        // DELETE plan is too simplistic for testing with `intermediateDistributedStage`
+        assertThatQuery(deleteQuery)
+                .withSetupQuery(setupQuery)
+                .withCleanupQuery(cleanupQuery)
+                .experiencing(TASK_FAILURE, Optional.of(ErrorType.INTERNAL_ERROR))
+                .at(intermediateDistributedStage())
+                .failsWithoutRetries(failure -> failure.hasMessageContaining(FAILURE_INJECTION_MESSAGE));
+
         assertThatQuery(deleteQuery)
                 .withSetupQuery(setupQuery)
                 .withCleanupQuery(cleanupQuery)
@@ -118,7 +104,7 @@ public abstract class BaseIcebergFailureRecoveryTest
                 .withCleanupQuery(cleanupQuery)
                 .experiencing(TASK_MANAGEMENT_REQUEST_TIMEOUT)
                 .at(boundaryDistributedStage())
-                .failsWithoutRetries(failure -> failure.hasMessageFindingMatch("Encountered too many errors talking to a worker node|Error closing remote buffer"))
+                .failsWithoutRetries(failure -> failure.hasMessageContaining("Encountered too many errors talking to a worker node"))
                 .finishesSuccessfully();
 
         if (getRetryPolicy() == RetryPolicy.QUERY) {
@@ -140,7 +126,6 @@ public abstract class BaseIcebergFailureRecoveryTest
         }
     }
 
-    // Copied from BaseDeltaFailureRecoveryTest
     @Override
     public void testUpdate()
     {
@@ -181,6 +166,14 @@ public abstract class BaseIcebergFailureRecoveryTest
                 .failsWithoutRetries(failure -> failure.hasMessageContaining(FAILURE_INJECTION_MESSAGE))
                 .finishesSuccessfully();
 
+        // UPDATE plan is too simplistic for testing with `intermediateDistributedStage`
+        assertThatQuery(updateQuery)
+                .withSetupQuery(setupQuery)
+                .withCleanupQuery(cleanupQuery)
+                .experiencing(TASK_FAILURE, Optional.of(ErrorType.INTERNAL_ERROR))
+                .at(intermediateDistributedStage())
+                .failsWithoutRetries(failure -> failure.hasMessageContaining(FAILURE_INJECTION_MESSAGE));
+
         assertThatQuery(updateQuery)
                 .withSetupQuery(setupQuery)
                 .withCleanupQuery(cleanupQuery)
@@ -194,7 +187,7 @@ public abstract class BaseIcebergFailureRecoveryTest
                 .withCleanupQuery(cleanupQuery)
                 .experiencing(TASK_MANAGEMENT_REQUEST_TIMEOUT)
                 .at(boundaryDistributedStage())
-                .failsWithoutRetries(failure -> failure.hasMessageFindingMatch("Encountered too many errors talking to a worker node|Error closing remote buffer"))
+                .failsWithoutRetries(failure -> failure.hasMessageContaining("Encountered too many errors talking to a worker node"))
                 .finishesSuccessfully();
 
         if (getRetryPolicy() == RetryPolicy.QUERY) {
@@ -216,11 +209,28 @@ public abstract class BaseIcebergFailureRecoveryTest
         }
     }
 
+    @Override
+    // materialized views are currently not implemented by Delta connector
+    public void testRefreshMaterializedView()
+    {
+        assertThatThrownBy(super::testRefreshMaterializedView)
+                .hasMessageContaining("This connector does not support creating materialized views");
+    }
+
+    @Test(invocationCount = INVOCATION_COUNT)
+    public void testCreatePartitionedTable()
+    {
+        testTableModification(
+                Optional.empty(),
+                "CREATE TABLE <table> WITH (partitioned_by = ARRAY['p']) AS SELECT *, 'partition1' p FROM orders",
+                Optional.of("DROP TABLE <table>"));
+    }
+
     @Test(invocationCount = INVOCATION_COUNT)
     public void testInsertIntoNewPartition()
     {
         testTableModification(
-                Optional.of("CREATE TABLE <table> WITH (partitioning = ARRAY['p']) AS SELECT *, 'partition1' p FROM orders"),
+                Optional.of("CREATE TABLE <table> WITH (partitioned_by = ARRAY['p']) AS SELECT *, 'partition1' p FROM orders"),
                 "INSERT INTO <table> SELECT *, 'partition2' p FROM orders",
                 Optional.of("DROP TABLE <table>"));
     }
@@ -229,25 +239,8 @@ public abstract class BaseIcebergFailureRecoveryTest
     public void testInsertIntoExistingPartition()
     {
         testTableModification(
-                Optional.of("CREATE TABLE <table> WITH (partitioning = ARRAY['p']) AS SELECT *, 'partition1' p FROM orders"),
+                Optional.of("CREATE TABLE <table> WITH (partitioned_by = ARRAY['p']) AS SELECT *, 'partition1' p FROM orders"),
                 "INSERT INTO <table> SELECT *, 'partition1' p FROM orders",
-                Optional.of("DROP TABLE <table>"));
-    }
-
-    @Test(invocationCount = INVOCATION_COUNT)
-    public void testMergePartitionedTable()
-    {
-        testTableModification(
-                Optional.of("CREATE TABLE <table> WITH (partitioning = ARRAY['bucket(orderkey, 10)']) AS SELECT * FROM orders"),
-                """
-                        MERGE INTO <table> t
-                        USING (SELECT orderkey, 'X' clerk FROM <table>) s
-                        ON t.orderkey = s.orderkey
-                        WHEN MATCHED AND s.orderkey > 1000
-                            THEN UPDATE SET clerk = t.clerk || s.clerk
-                        WHEN MATCHED AND s.orderkey <= 1000
-                            THEN DELETE
-                        """,
                 Optional.of("DROP TABLE <table>"));
     }
 }
