+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestHiveTaskFailureRecoveryTest.java
@@ -11,12 +11,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.faulttolerant.iceberg;
+package io.trino.plugin.hive;
 
+import com.google.common.collect.ImmutableMap;
 import io.trino.operator.RetryPolicy;
 import io.trino.plugin.exchange.filesystem.FileSystemExchangePlugin;
 import io.trino.plugin.exchange.filesystem.containers.MinioStorage;
-import io.trino.plugin.iceberg.IcebergQueryRunner;
+import io.trino.plugin.hive.containers.HiveMinioDataLake;
+import io.trino.plugin.hive.s3.S3HiveQueryRunner;
 import io.trino.testing.QueryRunner;
 import io.trino.tpch.TpchTable;
 import org.testng.annotations.AfterClass;
@@ -27,15 +29,16 @@ import java.util.Map;
 import static io.trino.plugin.exchange.filesystem.containers.MinioStorage.getExchangeManagerProperties;
 import static io.trino.testing.TestingNames.randomNameSuffix;
 
-public class TestIcebergQueryFailureRecoveryTest
-        extends BaseIcebergFailureRecoveryTest
+public class TestHiveTaskFailureRecoveryTest
+        extends BaseHiveFailureRecoveryTest
 {
-    private MinioStorage minioStorage;
-
-    protected TestIcebergQueryFailureRecoveryTest()
+    public TestHiveTaskFailureRecoveryTest()
     {
-        super(RetryPolicy.QUERY);
+        super(RetryPolicy.TASK);
     }
+
+    private HiveMinioDataLake hiveMinioDataLake;
+    private MinioStorage minioStorage;
 
     @Override
     protected QueryRunner createQueryRunner(
@@ -44,17 +47,26 @@ public class TestIcebergQueryFailureRecoveryTest
             Map<String, String> coordinatorProperties)
             throws Exception
     {
+        String bucketName = "test-hive-insert-overwrite-" + randomNameSuffix(); // randomizing bucket name to ensure cached TrinoS3FileSystem objects are not reused
+        this.hiveMinioDataLake = new HiveMinioDataLake(bucketName);
+        hiveMinioDataLake.start();
+
         this.minioStorage = new MinioStorage("test-exchange-spooling-" + randomNameSuffix());
         minioStorage.start();
 
-        return IcebergQueryRunner.builder()
+        return S3HiveQueryRunner.builder(hiveMinioDataLake)
                 .setInitialTables(requiredTpchTables)
-                .setCoordinatorProperties(coordinatorProperties)
                 .setExtraProperties(configProperties)
+                .setCoordinatorProperties(coordinatorProperties)
                 .setAdditionalSetup(runner -> {
                     runner.installPlugin(new FileSystemExchangePlugin());
                     runner.loadExchangeManager("filesystem", getExchangeManagerProperties(minioStorage));
                 })
+                .setHiveProperties(ImmutableMap.of(
+                        // Streaming upload allocates non trivial amount of memory for buffering (16MB per output file by default).
+                        // When streaming upload is enabled insert into a table with high number of buckets / partitions may cause
+                        // the tests to run out of memory as the buffer space is eagerly allocated for each output file.
+                        "hive.s3.streaming.enabled", "false"))
                 .build();
     }
 
@@ -62,6 +74,10 @@ public class TestIcebergQueryFailureRecoveryTest
     public void destroy()
             throws Exception
     {
+        if (hiveMinioDataLake != null) {
+            hiveMinioDataLake.close();
+            hiveMinioDataLake = null;
+        }
         if (minioStorage != null) {
             minioStorage.close();
             minioStorage = null;
