+++ plugin/trino-iceberg/src/test/java/io/trino/plugin/iceberg/TestIcebergQueryFailureRecoveryTest.java
@@ -11,14 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.faulttolerant.hive;
+package io.trino.plugin.iceberg;
 
-import com.google.common.collect.ImmutableMap;
 import io.trino.operator.RetryPolicy;
 import io.trino.plugin.exchange.filesystem.FileSystemExchangePlugin;
 import io.trino.plugin.exchange.filesystem.containers.MinioStorage;
-import io.trino.plugin.hive.containers.HiveMinioDataLake;
-import io.trino.plugin.hive.s3.S3HiveQueryRunner;
 import io.trino.testing.QueryRunner;
 import io.trino.tpch.TpchTable;
 import org.testng.annotations.AfterClass;
@@ -29,16 +26,15 @@ import java.util.Map;
 import static io.trino.plugin.exchange.filesystem.containers.MinioStorage.getExchangeManagerProperties;
 import static io.trino.testing.TestingNames.randomNameSuffix;
 
-public class TestHiveQueryFailureRecoveryTest
-        extends BaseHiveFailureRecoveryTest
+public class TestIcebergQueryFailureRecoveryTest
+        extends BaseIcebergFailureRecoveryTest
 {
-    public TestHiveQueryFailureRecoveryTest()
+    private MinioStorage minioStorage;
+
+    protected TestIcebergQueryFailureRecoveryTest()
     {
         super(RetryPolicy.QUERY);
     }
-
-    private HiveMinioDataLake hiveMinioDataLake;
-    private MinioStorage minioStorage;
 
     @Override
     protected QueryRunner createQueryRunner(
@@ -47,26 +43,17 @@ public class TestHiveQueryFailureRecoveryTest
             Map<String, String> coordinatorProperties)
             throws Exception
     {
-        String bucketName = "test-hive-insert-overwrite-" + randomNameSuffix(); // randomizing bucket name to ensure cached TrinoS3FileSystem objects are not reused
-        this.hiveMinioDataLake = new HiveMinioDataLake(bucketName);
-        hiveMinioDataLake.start();
-
         this.minioStorage = new MinioStorage("test-exchange-spooling-" + randomNameSuffix());
         minioStorage.start();
 
-        return S3HiveQueryRunner.builder(hiveMinioDataLake)
+        return IcebergQueryRunner.builder()
                 .setInitialTables(requiredTpchTables)
-                .setExtraProperties(configProperties)
                 .setCoordinatorProperties(coordinatorProperties)
+                .setExtraProperties(configProperties)
                 .setAdditionalSetup(runner -> {
                     runner.installPlugin(new FileSystemExchangePlugin());
                     runner.loadExchangeManager("filesystem", getExchangeManagerProperties(minioStorage));
                 })
-                .setHiveProperties(ImmutableMap.of(
-                        // Streaming upload allocates non trivial amount of memory for buffering (16MB per output file by default).
-                        // When streaming upload is enabled insert into a table with high number of buckets / partitions may cause
-                        // the tests to run out of memory as the buffer space is eagerly allocated for each output file.
-                        "hive.s3.streaming.enabled", "false"))
                 .build();
     }
 
@@ -74,10 +61,6 @@ public class TestHiveQueryFailureRecoveryTest
     public void destroy()
             throws Exception
     {
-        if (hiveMinioDataLake != null) {
-            hiveMinioDataLake.close();
-            hiveMinioDataLake = null;
-        }
         if (minioStorage != null) {
             minioStorage.close();
             minioStorage = null;
