+++ testing/trino-faulttolerant-tests/pom.xml
@@ -490,10 +490,6 @@
                                 <exclude>**/io/trino/faulttolerant/delta/Test*.java</exclude>
                                 <exclude>**/io/trino/faulttolerant/hive/Test*.java</exclude>
                                 <exclude>**/io/trino/faulttolerant/iceberg/Test*.java</exclude>
-                                <exclude>**/io/trino/faulttolerant/mongodb/Test*.java</exclude>
-                                <exclude>**/io/trino/faulttolerant/mysql/Test*.java</exclude>
-                                <exclude>**/io/trino/faulttolerant/postgresql/Test*.java</exclude>
-                                <exclude>**/io/trino/faulttolerant/sqlserver/Test*.java</exclude>
                             </excludes>
                         </configuration>
                     </plugin>
@@ -502,26 +498,7 @@
         </profile>
 
         <profile>
-            <id>test-fault-tolerant-hive-1</id>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <configuration>
-                            <!-- Failure recovery tests spend most of the time waiting for a retry -->
-                            <threadCount>4</threadCount>
-                            <includes>
-                                <include>**/io/trino/faulttolerant/hive/*FailureRecoveryTest.java</include>
-                            </includes>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-
-        <profile>
-            <id>test-fault-tolerant-hive-2</id>
+            <id>test-fault-tolerant-hive</id>
             <build>
                 <plugins>
                     <plugin>
@@ -570,78 +547,6 @@
                             <threadCount>4</threadCount>
                             <includes>
                                 <include>**/io/trino/faulttolerant/iceberg/Test*.java</include>
-                            </includes>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-
-        <profile>
-            <id>test-fault-tolerant-postgresql</id>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <configuration>
-                            <threadCount>4</threadCount>
-                            <includes>
-                                <include>**/io/trino/faulttolerant/postgresql/Test*.java</include>
-                            </includes>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-
-        <profile>
-            <id>test-fault-tolerant-mysql</id>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <configuration>
-                            <threadCount>4</threadCount>
-                            <includes>
-                                <include>**/io/trino/faulttolerant/mysql/Test*.java</include>
-                            </includes>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-
-        <profile>
-            <id>test-fault-tolerant-sqlserver</id>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <configuration>
-                            <threadCount>4</threadCount>
-                            <includes>
-                                <include>**/io/trino/faulttolerant/sqlserver/Test*.java</include>
-                            </includes>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-
-        <profile>
-            <id>test-fault-tolerant-mongodb</id>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <configuration>
-                            <threadCount>4</threadCount>
-                            <includes>
-                                <include>**/io/trino/faulttolerant/mongodb/Test*.java</include>
                             </includes>
                         </configuration>
                     </plugin>
